modify/app/MedicalBot/requirements.txt
aiohttp
botbuilder-integration-aiohttp>=4.11.0
requests==2.25.1